Add an Expense to the Data Frame

This function adds a new expense entry to the given data frame of expenses. It checks for the validity of the category and amount before adding.

add_expense(expenses_data, category, amount)

Arguments

expenses_data

A data frame containing existing expenses.

category

The category of the new expense.

amount

The amount of the new expense.

Value

A data frame with the new expense added.

Author

Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.

Examples

expenses <- data.frame(category = character(), amount = numeric())
expenses <- add_expense(expenses, "Food", 100)