Create a Bar Chart with Plotly

This function creates an interactive bar chart using ggplot2 and Plotly. It is designed to visualize data in categories with their corresponding amounts. The chart uses a minimalistic color palette and hides the legend for a cleaner appearance.

createBarChart(
  data,
  title = "Income, Expenses, and Savings",
  xLabel = "Category",
  yLabel = "Amount"
)

Arguments

data

A data frame containing the variables `category` and `amount`.

title

Title of the plot, default is "Income, Expenses, and Savings".

xLabel

Label for the x-axis, default is "Category".

yLabel

Label for the y-axis, default is "Amount".

Value

An interactive Plotly object.

Author

Group C composed of Marc Bourleau, Eleonore Gillain, Khrystyna Khmilovska and Konstantinos Kourlimpinis.

Examples

test_data <- data.frame(category = c("Income", "Expenses", "Savings"),
                       amount = c(1000, 500, 500))
createBarChart(test_data)
#> Loading required package: ggplot2
#> Warning: package 'ggplot2' was built under R version 4.2.3
#> Loading required package: plotly
#> Warning: package 'plotly' was built under R version 4.2.2
#> 
#> Attaching package: 'plotly'
#> The following object is masked from 'package:ggplot2':
#> 
#>     last_plot
#> The following object is masked from 'package:stats':
#> 
#>     filter
#> The following object is masked from 'package:graphics':
#> 
#>     layout